requirements.txt
PyJWT~=2.4.0
Flask~=2.0.3